+++ single.php
@@ -18,12 +18,14 @@
 			<div class="post">
 		
 				<h1 class="headline remove_bottom"><?php the_title(); ?></h1>
-				<p class="postmeta half_bottom">
-					by <a href="<?php echo get_author_posts_url( get_the_author_meta( 'ID' ) ); ?>"><?php the_author_meta( 'display_name' ); ?></a> on <?php the_time('F jS, Y'); ?>
-					<?php if ( is_user_logged_in() ) { edit_post_link( 'edit post', ' [', ']' ); } ?>
-				</p>
 				
-				<?php if ( has_post_thumbnail() && ( has_tag('big-image') || has_tag('big-image-everywhere') ) ) {
+				<div class="postmeta">
+					<div class="comment_link th_comment_link"><a href="#disqus_thread"><div class="comment_bubble"></div> <?php comments_number('leave a comment','1 comment','% comments'); ?></a></div>
+					<div class="author_link">by <?php the_author(); ?></div>
+					<div class="clear"></div>
+				</div>
+
+				<?php if ( has_post_thumbnail() && has_tag('big-image') ) {
 					the_post_thumbnail('large');
 				}